requirements.txt
Django == 4.1.1
Pillow == 9.2.0
django-crispy-forms
django-recaptcha
requests